+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Changelog
 
+### 1.0.1, 2021-11-08
+
+* changed: minimum Gravity Forms version is 2.4.18
+* added: entry note to show that this add-on marked event as spam
+
 ### 1.0.0, 2021-11-08
 
 * initial release (extracted from a larger codebase)